refactor: change tests structure (#115)
