+++ src/Shared/Domain/Tests/Unit/ValueObject/Date/TimeTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Shared\Tests\Unit\ValueObject\Date;
+namespace App\Shared\Domain\Tests\Unit\ValueObject\Date;
 
 use App\Shared\Domain\Exception\InvalidArgumentException;
 use App\Shared\Domain\ValueObject\Date\DateTime;
