+++ src/Shared/Domain/Tests/Unit/ValueObject/Token/UuidTokenTest.php
@@ -2,26 +2,26 @@
 
 declare(strict_types=1);
 
-namespace App\Shared\Tests\Unit\ValueObject\Identifier;
+namespace App\Shared\Domain\Tests\Unit\ValueObject\Token;
 
 use App\Shared\Domain\Exception\InvalidArgumentException;
-use App\Shared\Domain\ValueObject\Identifier\UuidIdentifier;
+use App\Shared\Domain\ValueObject\Token\UuidToken;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Uid\Uuid;
 
-final class UuidIdentifierTest extends TestCase
+final class UuidTokenTest extends TestCase
 {
-    public function testIfFactoriesCreateIdentifier(): void
+    public function testIfFactoriesCreateToken(): void
     {
-        $identifier = UuidIdentifier::create();
+        $identifier = UuidToken::create();
         $this->assertTrue(Uuid::isValid((string) $identifier));
         $this->assertInstanceOf(Uuid::class, $identifier->uuid());
 
-        $identifier = UuidIdentifier::createFromString((string) Uuid::v4());
+        $identifier = UuidToken::createFromString((string) Uuid::v4());
         $this->assertTrue(Uuid::isValid((string) $identifier));
         $this->assertInstanceOf(Uuid::class, $identifier->uuid());
 
-        $identifier = UuidIdentifier::createFromUuid(Uuid::v4());
+        $identifier = UuidToken::createFromUuid(Uuid::v4());
         $this->assertTrue(Uuid::isValid((string) $identifier));
         $this->assertInstanceOf(Uuid::class, $identifier->uuid());
     }
@@ -29,6 +29,6 @@ final class UuidIdentifierTest extends TestCase
     public function testIfUuidIsInvalid(): void
     {
         $this->expectException(InvalidArgumentException::class);
-        UuidIdentifier::createFromString('fail');
+        UuidToken::createFromString('fail');
     }
 }
